DOM: Observable subscription uses Callback or Dictionary union
This CL addresses https://github.com/WICG/observable/issues/71, which
calls for an ergonomic design improvement whereby you can subscribe with
either a full-fledged `Observer` dictionary, or just a single callback
which is used as the subscription's next handler.

Subscribing with a single callback is identical to subscribing with an
`Observer` dictionary whose `next` handler is provided and nothing else.

R=masonf@chromium.org

Bug: 1485981
Change-Id: I5781a6cf404d97ac2de2bc88a303bcd522f63ec5