Fix progress function calculation
We can simplify the progress function when either all the operands
resolve to the same canonical unit or all the operands have the same type.
Currently `progress(10em from 10em to 100em)` crashes as we try to compute the canonical value in either case of simplification.

This CL changes this behavior to distinguish which simplification path
to use, so that in the mentioned progress() example we just take the
double values when all the operands are of the same type.

Bug: 1503730
Change-Id: I361477ba2a03fd6063861d1d1a82d4b77fa77f2f