[FLEDGE]: Add allSlotsRequestedSizes to AuctionConfig.
This new optional field will be sent to InterestGroups that request it
when requesting trustedBiddingSignals.

This CL adds some WPT tests for requestedSize and
allSlotsRequestedSizes in the parse failure case, but no tests for
the successful parse case.  I'll add those once the feature does
something.

This also introduces a new feature to enable/disable the new
argument.

Bug: 1506238
Change-Id: I9c7706c69f8bbfafd52fcefadc390c70174a762f