compute-pressure: Ensure clean test exit
It can happen on certain bots that the pressure updates
are still sent before disconnect() is processed in the
mockPressureService, which can lead in certain cases to
receive more updates than allowed in the test.

This patch add a `gotPenalty` flag, to make sure that at least
one penalty was applied when reaching the maximum allowed updates.

Change-Id: I535b8c21370d14f13f0d6140d1eb2cc043cbeac2
Bug: 1501324