[soft navigations] Ignore non-attributed LCP before softnav detection
This CL ignores LCPs that are not attribution to the soft navigation
task after a user interaction and before a soft navigation was
detected, as discussed in [1].

It also fixes a bug in soft navigation detection in the case where all
paints arrive before the URL was changed. Currently in such cases, a
soft navigation entry is not emitted. This CL fixes that.

[1] https://github.com/WICG/soft-navigations/issues/29#issuecomment-1831386977

Bug: 1505994
Change-Id: I444ff03c0b64663b913b4e73a7c820b7bab11047