[css-images-4] Implement unprefixed cross-fade().
This is similar to -webkit-cross-fade(), but differs in that it
has a different syntax, and supports more (or less) than two images.
It also supports cross-fading against a solid color, which we implement
internally as a gradient for simplicity. (This may not be optimal for
memory or performance, but it should do for now.)

We do not implement parts of the spec that are not yet written
but will be, in particular simplification at computed value time.

The functionality is behind a new runtime flag CSSCrossFade,
which is set to experimental as usual.

Change-Id: I4d29e14b215a65223b98dcaee8ba99b8f62d94dc